+++ lib/src/border/bubble_border.dart
@@ -43,6 +43,112 @@ class BubbleBorder extends OutlinedBorder {
     );
   }
 
+  Path _getPath(RRect original) {
+    final helfStroke = side.width / 2.0;
+    final rrect = original.deflate(helfStroke);
+    final rect = rrect.outerRect;
+
+    final ratio = rect.size.width / original.outerRect.size.width;
+    final resizedArrowWidth = arrowWidth * ratio;
+    final range = (rect.width - resizedArrowWidth) / 2.0;
+
+    final halfArrowWidth = resizedArrowWidth / 2.0;
+    final resizedArrowHeight = arrowHeight;
+
+    switch (direction) {
+      case ShapeDirection.up:
+        final minOffset = max(range - rrect.blRadiusX, 0);
+        final maxOffset = max(range - rrect.brRadiusX, 0);
+        final offset = arrowOffset.clamp(-minOffset, maxOffset);
+        return Path()
+          ..moveTo(rect.bottomCenter.dx + offset, rect.bottomCenter.dy)
+          ..lineTo(rect.bottomCenter.dx - halfArrowWidth + offset, rect.bottomCenter.dy - resizedArrowHeight)
+          ..lineTo(rect.bottomLeft.dx + rrect.blRadiusX, rect.bottomLeft.dy - resizedArrowHeight)
+          ..quadraticBezierTo(rect.bottomLeft.dx, rect.bottomLeft.dy - resizedArrowHeight, rect.bottomLeft.dx,
+              rect.bottomLeft.dy - rrect.blRadiusY - resizedArrowHeight)
+          ..lineTo(rect.topLeft.dx, rect.topLeft.dy + rrect.tlRadiusY)
+          ..quadraticBezierTo(rect.topLeft.dx, rect.topLeft.dy, rect.topLeft.dx + rrect.tlRadiusX, rect.topLeft.dy)
+          ..lineTo(rect.topRight.dx - rrect.trRadiusX, rect.topRight.dy)
+          ..quadraticBezierTo(rect.topRight.dx, rect.topRight.dy, rect.topRight.dx, rect.topRight.dy + rrect.trRadiusY)
+          ..lineTo(rect.bottomRight.dx, rect.bottomRight.dy - rrect.brRadiusY - resizedArrowHeight)
+          ..quadraticBezierTo(rect.bottomRight.dx, rect.bottomRight.dy - resizedArrowHeight,
+              rect.bottomRight.dx - rrect.brRadiusX, rect.bottomRight.dy - resizedArrowHeight)
+          ..lineTo(rect.bottomCenter.dx + halfArrowWidth + offset, rect.bottomCenter.dy - resizedArrowHeight)
+          ..close();
+      case ShapeDirection.down:
+        final minOffset = max(range - rrect.tlRadiusX, 0);
+        final maxOffset = max(range - rrect.trRadiusX, 0);
+        final offset = arrowOffset.clamp(-minOffset, maxOffset);
+        return Path()
+          ..moveTo(rect.topCenter.dx + offset, rect.topCenter.dy)
+          ..lineTo(rect.topCenter.dx + halfArrowWidth + offset, rect.topCenter.dy + resizedArrowHeight)
+          ..lineTo(rect.topRight.dx - rrect.trRadiusX, rect.topRight.dy + resizedArrowHeight)
+          ..quadraticBezierTo(rect.topRight.dx, rect.topRight.dy + resizedArrowHeight, rect.topRight.dx,
+              rect.topRight.dy + rrect.trRadiusY + resizedArrowHeight)
+          ..lineTo(rect.bottomRight.dx, rect.bottomRight.dy - rrect.brRadiusY)
+          ..quadraticBezierTo(
+              rect.bottomRight.dx, rect.bottomRight.dy, rect.bottomRight.dx - rrect.brRadiusX, rect.bottomRight.dy)
+          ..lineTo(rect.bottomLeft.dx + rrect.blRadiusX, rect.bottomLeft.dy)
+          ..quadraticBezierTo(
+              rect.bottomLeft.dx, rect.bottomLeft.dy, rect.bottomLeft.dx, rect.bottomLeft.dy - rrect.blRadiusY)
+          ..lineTo(rect.topLeft.dx, rect.topLeft.dy + rrect.tlRadiusY + resizedArrowHeight)
+          ..quadraticBezierTo(rect.topLeft.dx, rect.topLeft.dy + resizedArrowHeight, rect.topLeft.dx + rrect.tlRadiusX,
+              rect.topLeft.dy + resizedArrowHeight)
+          ..lineTo(rect.topCenter.dx - halfArrowWidth + offset, rect.topCenter.dy + resizedArrowHeight)
+          ..close();
+      case ShapeDirection.left:
+        final minOffset = max(range - rrect.tlRadiusY, 0);
+        final maxOffset = max(range - rrect.blRadiusY, 0);
+        final offset = arrowOffset.clamp(-minOffset, maxOffset);
+        return Path()
+          ..moveTo(rect.centerRight.dx, rect.centerRight.dy + offset)
+          ..lineTo(rect.centerRight.dx - resizedArrowHeight, rect.centerRight.dy + halfArrowWidth + offset)
+          ..lineTo(rect.bottomRight.dx - resizedArrowHeight, rect.bottomRight.dy - rrect.brRadiusY)
+          ..quadraticBezierTo(rect.bottomRight.dx - resizedArrowHeight, rect.bottomRight.dy,
+              rect.bottomRight.dx - rrect.brRadiusX - resizedArrowHeight, rect.bottomRight.dy)
+          ..lineTo(rect.bottomLeft.dx + rrect.blRadiusX, rect.bottomLeft.dy)
+          ..quadraticBezierTo(
+              rect.bottomLeft.dx, rect.bottomLeft.dy, rect.bottomLeft.dx, rect.bottomLeft.dy - rrect.blRadiusY)
+          ..lineTo(rect.topLeft.dx, rect.topLeft.dy + rrect.tlRadiusY)
+          ..quadraticBezierTo(rect.topLeft.dx, rect.topLeft.dy, rect.topLeft.dx + rrect.tlRadiusX, rect.topLeft.dy)
+          ..lineTo(rect.topRight.dx - resizedArrowHeight - rrect.trRadiusX, rect.topRight.dy)
+          ..quadraticBezierTo(rect.topRight.dx - resizedArrowHeight, rect.topRight.dy,
+              rect.topRight.dx - resizedArrowHeight, rect.topRight.dy + rrect.trRadiusY)
+          ..lineTo(rect.centerRight.dx - resizedArrowHeight, rect.centerRight.dy - halfArrowWidth + offset)
+          ..close();
+      case ShapeDirection.right:
+        final minOffset = max(range - rrect.trRadiusY, 0);
+        final maxOffset = max(range - rrect.brRadiusY, 0);
+        final offset = arrowOffset.clamp(-minOffset, maxOffset);
+        return Path()
+          ..moveTo(rect.centerLeft.dx, rect.centerLeft.dy + offset)
+          ..lineTo(rect.centerLeft.dx + resizedArrowHeight, rect.centerLeft.dy - halfArrowWidth + offset)
+          ..lineTo(rect.topLeft.dx + resizedArrowHeight, rect.topLeft.dy + rrect.tlRadiusY)
+          ..quadraticBezierTo(rect.topLeft.dx + resizedArrowHeight, rect.topLeft.dy,
+              rect.topLeft.dx + rrect.tlRadiusX + resizedArrowHeight, rect.topLeft.dy)
+          ..lineTo(rect.topRight.dx - rrect.trRadiusX, rect.topRight.dy)
+          ..quadraticBezierTo(rect.topRight.dx, rect.topRight.dy, rect.topRight.dx, rect.topRight.dy + rrect.trRadiusY)
+          ..lineTo(rect.bottomRight.dx, rect.bottomRight.dy - rrect.brRadiusY)
+          ..quadraticBezierTo(
+              rect.bottomRight.dx, rect.bottomRight.dy, rect.bottomRight.dx - rrect.brRadiusX, rect.bottomRight.dy)
+          ..lineTo(rect.bottomLeft.dx + resizedArrowHeight + rrect.blRadiusX, rect.bottomLeft.dy)
+          ..quadraticBezierTo(rect.bottomLeft.dx + resizedArrowHeight, rect.bottomLeft.dy,
+              rect.bottomLeft.dx + resizedArrowHeight, rect.bottomLeft.dy - rrect.blRadiusY)
+          ..lineTo(rect.centerLeft.dx + resizedArrowHeight, rect.centerLeft.dy + halfArrowWidth + offset)
+          ..close();
+    }
+  }
+
+  @override
+  Path getInnerPath(Rect rect, {TextDirection? textDirection}) {
+    return _getPath(borderRadius.resolve(textDirection).toRRect(rect).deflate(side.width));
+  }
+
+  @override
+  Path getOuterPath(Rect rect, {TextDirection? textDirection}) {
+    return _getPath(borderRadius.resolve(textDirection).toRRect(rect));
+  }
+
   @override
   BubbleBorder copyWith({
     ShapeDirection? direction,
@@ -63,24 +169,17 @@ class BubbleBorder extends OutlinedBorder {
   }
 
   @override
-  Path getInnerPath(Rect rect, {TextDirection? textDirection}) {
-    return _bubblePath(rect.deflate(side.width));
-  }
-
-  @override
-  Path getOuterPath(Rect rect, {TextDirection? textDirection}) {
-    return _bubblePath(rect);
-  }
-
-  @override
   void paint(Canvas canvas, Rect rect, {TextDirection? textDirection}) {
     switch (side.style) {
       case BorderStyle.none:
         break;
       case BorderStyle.solid:
-        final paint = side.toPaint();
-        paint.strokeWidth = side.width * 2;
-        canvas.drawPath(_bubblePath(rect), paint);
+        canvas.drawPath(
+          getOuterPath(rect, textDirection: textDirection),
+          side.toPaint()
+            ..strokeJoin = StrokeJoin.round
+            ..strokeCap = StrokeCap.round,
+        );
     }
   }
 
@@ -102,96 +201,5 @@ class BubbleBorder extends OutlinedBorder {
   @override
   String toString() {
     return '${objectRuntimeType(this, 'BubbleBorder')}($side)';
-  }
-
-  Path _bubblePath(Rect rect) {
-    final range = (rect.width - arrowWidth) / 2.0;
-
-    switch (direction) {
-      case ShapeDirection.up:
-        final minOffset = max(range - borderRadius.bottomLeft.x, 0);
-        final maxOffset = max(range - borderRadius.bottomRight.x, 0);
-        final offset = arrowOffset.clamp(-minOffset, maxOffset);
-        return Path()
-          ..moveTo(rect.bottomCenter.dx + offset, rect.bottomCenter.dy)
-          ..lineTo(rect.bottomCenter.dx - (arrowWidth / 2.0) + offset, rect.bottomCenter.dy - arrowHeight)
-          ..lineTo(rect.bottomLeft.dx + borderRadius.bottomLeft.x, rect.bottomLeft.dy - arrowHeight)
-          ..quadraticBezierTo(rect.bottomLeft.dx, rect.bottomLeft.dy - arrowHeight, rect.bottomLeft.dx,
-              rect.bottomLeft.dy - borderRadius.bottomLeft.y - arrowHeight)
-          ..lineTo(rect.topLeft.dx, rect.topLeft.dy + borderRadius.topLeft.y)
-          ..quadraticBezierTo(
-              rect.topLeft.dx, rect.topLeft.dy, rect.topLeft.dx + borderRadius.topLeft.x, rect.topLeft.dy)
-          ..lineTo(rect.topRight.dx - borderRadius.topRight.x, rect.topRight.dy)
-          ..quadraticBezierTo(
-              rect.topRight.dx, rect.topRight.dy, rect.topRight.dx, rect.topRight.dy + borderRadius.topRight.y)
-          ..lineTo(rect.bottomRight.dx, rect.bottomRight.dy - borderRadius.bottomRight.y - arrowHeight)
-          ..quadraticBezierTo(rect.bottomRight.dx, rect.bottomRight.dy - arrowHeight,
-              rect.bottomRight.dx - borderRadius.bottomRight.x, rect.bottomRight.dy - arrowHeight)
-          ..lineTo(rect.bottomCenter.dx + (arrowWidth / 2.0) + offset, rect.bottomCenter.dy - arrowHeight)
-          ..close();
-      case ShapeDirection.down:
-        final minOffset = max(range - borderRadius.topLeft.x, 0);
-        final maxOffset = max(range - borderRadius.topRight.x, 0);
-        final offset = arrowOffset.clamp(-minOffset, maxOffset);
-        return Path()
-          ..moveTo(rect.topCenter.dx + offset, rect.topCenter.dy)
-          ..lineTo(rect.topCenter.dx + (arrowWidth / 2.0) + offset, rect.topCenter.dy + arrowHeight)
-          ..lineTo(rect.topRight.dx - borderRadius.topRight.x, rect.topRight.dy + arrowHeight)
-          ..quadraticBezierTo(rect.topRight.dx, rect.topRight.dy + arrowHeight, rect.topRight.dx,
-              rect.topRight.dy + borderRadius.topRight.y + arrowHeight)
-          ..lineTo(rect.bottomRight.dx, rect.bottomRight.dy - borderRadius.bottomRight.y)
-          ..quadraticBezierTo(rect.bottomRight.dx, rect.bottomRight.dy,
-              rect.bottomRight.dx - borderRadius.bottomRight.x, rect.bottomRight.dy)
-          ..lineTo(rect.bottomLeft.dx + borderRadius.bottomLeft.x, rect.bottomLeft.dy)
-          ..quadraticBezierTo(rect.bottomLeft.dx, rect.bottomLeft.dy, rect.bottomLeft.dx,
-              rect.bottomLeft.dy - borderRadius.bottomLeft.y)
-          ..lineTo(rect.topLeft.dx, rect.topLeft.dy + borderRadius.topLeft.y + arrowHeight)
-          ..quadraticBezierTo(rect.topLeft.dx, rect.topLeft.dy + arrowHeight, rect.topLeft.dx + borderRadius.topLeft.x,
-              rect.topLeft.dy + arrowHeight)
-          ..lineTo(rect.topCenter.dx - (arrowWidth / 2.0) + offset, rect.topCenter.dy + arrowHeight)
-          ..close();
-      case ShapeDirection.left:
-        final minOffset = max(range - borderRadius.topLeft.y, 0);
-        final maxOffset = max(range - borderRadius.bottomLeft.y, 0);
-        final offset = arrowOffset.clamp(-minOffset, maxOffset);
-        return Path()
-          ..moveTo(rect.centerRight.dx, rect.centerRight.dy + offset)
-          ..lineTo(rect.centerRight.dx - arrowHeight, rect.centerRight.dy + (arrowWidth / 2.0) + offset)
-          ..lineTo(rect.bottomRight.dx - arrowHeight, rect.bottomRight.dy - borderRadius.bottomRight.y)
-          ..quadraticBezierTo(rect.bottomRight.dx - arrowHeight, rect.bottomRight.dy,
-              rect.bottomRight.dx - borderRadius.bottomRight.x - arrowHeight, rect.bottomRight.dy)
-          ..lineTo(rect.bottomLeft.dx + borderRadius.bottomLeft.x, rect.bottomLeft.dy)
-          ..quadraticBezierTo(rect.bottomLeft.dx, rect.bottomLeft.dy, rect.bottomLeft.dx,
-              rect.bottomLeft.dy - borderRadius.bottomLeft.y)
-          ..lineTo(rect.topLeft.dx, rect.topLeft.dy + borderRadius.topLeft.y)
-          ..quadraticBezierTo(
-              rect.topLeft.dx, rect.topLeft.dy, rect.topLeft.dx + borderRadius.topLeft.x, rect.topLeft.dy)
-          ..lineTo(rect.topRight.dx - arrowHeight - borderRadius.topRight.x, rect.topRight.dy)
-          ..quadraticBezierTo(rect.topRight.dx - arrowHeight, rect.topRight.dy, rect.topRight.dx - arrowHeight,
-              rect.topRight.dy + borderRadius.topRight.y)
-          ..lineTo(rect.centerRight.dx - arrowHeight, rect.centerRight.dy - (arrowWidth / 2.0) + offset)
-          ..close();
-      case ShapeDirection.right:
-        final minOffset = max(range - borderRadius.topRight.y, 0);
-        final maxOffset = max(range - borderRadius.bottomRight.y, 0);
-        final offset = arrowOffset.clamp(-minOffset, maxOffset);
-        return Path()
-          ..moveTo(rect.centerLeft.dx, rect.centerLeft.dy + offset)
-          ..lineTo(rect.centerLeft.dx + arrowHeight, rect.centerLeft.dy - (arrowWidth / 2.0) + offset)
-          ..lineTo(rect.topLeft.dx + arrowHeight, rect.topLeft.dy + borderRadius.topLeft.y)
-          ..quadraticBezierTo(rect.topLeft.dx + arrowHeight, rect.topLeft.dy,
-              rect.topLeft.dx + borderRadius.topLeft.x + arrowHeight, rect.topLeft.dy)
-          ..lineTo(rect.topRight.dx - borderRadius.topRight.x, rect.topRight.dy)
-          ..quadraticBezierTo(
-              rect.topRight.dx, rect.topRight.dy, rect.topRight.dx, rect.topRight.dy + borderRadius.topRight.y)
-          ..lineTo(rect.bottomRight.dx, rect.bottomRight.dy - borderRadius.bottomRight.y)
-          ..quadraticBezierTo(rect.bottomRight.dx, rect.bottomRight.dy,
-              rect.bottomRight.dx - borderRadius.bottomRight.x, rect.bottomRight.dy)
-          ..lineTo(rect.bottomLeft.dx + arrowHeight + borderRadius.bottomLeft.x, rect.bottomLeft.dy)
-          ..quadraticBezierTo(rect.bottomLeft.dx + arrowHeight, rect.bottomLeft.dy, rect.bottomLeft.dx + arrowHeight,
-              rect.bottomLeft.dy - borderRadius.bottomLeft.y)
-          ..lineTo(rect.centerLeft.dx + arrowHeight, rect.centerLeft.dy + (arrowWidth / 2.0) + offset)
-          ..close();
-    }
   }
 }