+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 1.0.0
+## 2.0.0
 
-* Initial release.
+* Implements the Material and Cupertino styled buttons using platform widgets.
+* Change the usage of ShapedButtonStyle to distinguish the behavior on different platform.
+* Updated the BubbleBorder to draw a more accurate path.